WIP - disabling input for buyer/vendor if their contract hasn't arrived in the resolve dispute form
