+++ js/languages/en_US.json
@@ -1618,6 +1618,8 @@
       "heading": "Resolve Dispute",
       "buyerAmountLabel": "Buyer Amount",
       "vendorAmountLabel": "Vendor Amount",
+      "vendorContractUnavailable": "You cannot award any funds to the vendor until their contract arrives. It will be sent over the next time they come online.",
+      "buyerContractUnavailable": "You cannot award any funds to the buyer until their contract arrives. It will be sent over the next time they come online.",
       "commentLabel": "Comment",
       "commentPlaceholder": "Explain your decision...",
       "btnCancel": "Cancel",
