--- js/templates/modals/orderDetail/contract.html
@@ -1,51 +1,0 @@
-<div class="padLg flexVCent">
-  <div class="backToSummaryWrap">
-    <a class="js-backToSummary clrTEm txU"><%= ob.polyT(`orderDetail.backToSummary`) %></a>
-  </div>
-  <div class="txCtr txB tx3 flexExpand"><%= ob.polyT(`orderDetail.contractTab.heading`) %></div>
-</div>
-<hr class="clrBr rowMd" />
-<% if (!ob.isCase) { %>
-  <div class="border clrBr clrP clrT js-jsonContractContainer"></div>
-<% } else { %>
-  <% if (!ob.otherContractAvailable) {
-    const unavailableErrKey =
-      `orderDetail.contractTab.${ob.buyerOpened ? 'vendor' : 'buyer'}ContractNotAvailable`;
-    const keyPreface = ob.buyerOpened ? 'buyer' : 'vendor';
-  %>
-    <p class="rowLg"><%= ob.polyT(unavailableErrKey) %></p>
-
-    <div class="tx4 txB row"><%= ob.polyT(`orderDetail.contractTab.${keyPreface}ContractHeading`) %></div>
-    <% if (!ob[`${keyPreface}ContractVerified`]) { %>
-      <% // only available contract has failed verification %>
-      <p class="clrTErr flexVCent"><span class="ion-ios-close-empty tx1 margRSm"></span><%= ob.polyT(`orderDetail.contractTab.${keyPreface}ContractNotAuthentic`) %></p>    
-    <% } %>
-    <div class="border clrBr clrP clrT <% print(`js-json${ob.capitalize(keyPreface)}ContractContainer`) %>"></div>    
-  <% } else if (ob.buyerContractVerified && ob.vendorContractVerified) { %>
-    <% // if both contracts are authentic, we'll only show the one we're guaranteed to have %>
-    <p class="clrTEm flexVCent"><span class="ion-ios-checkmark-outline tx1 margRSm"></span><%= ob.polyT('orderDetail.contractTab.buyerVendorContractVerified') %></p>
-    <div class="border clrBr clrP clrT <% print(`js-json${ob.buyerOpened ? 'Buyer' : 'Vendor'}ContractContainer`) %>"></div>
-  <% } else if (!ob.buyerContractVerified && !ob.vendorContractVerified) { %>
-    <% // both contracts are forged %>
-    <div class="tx4 txB row"><%= ob.polyT('orderDetail.contractTab.buyerContractHeading') %></div>
-    <p class="clrTErr flexVCent"><span class="ion-ios-close-empty tx1 margRSm"></span><%= ob.polyT('orderDetail.contractTab.buyerContractNotAuthentic') %></p>    
-    <div class="border clrBr clrP clrT rowLg js-jsonBuyerContractContainer"></div>
-    <div class="tx4 txB row"><%= ob.polyT('orderDetail.contractTab.vendorContractHeading') %></div>
-    <p class="clrTErr flexVCent"><span class="ion-ios-close-empty tx1 margRSm"></span><%= ob.polyT('orderDetail.contractTab.vendorContractNotAuthentic') %></p>    
-    <div class="border clrBr clrP clrT js-jsonVendorContractContainer"></div>
-  <% } else if (!ob.buyerContractVerified) { %>
-    <% // buyers contract is forged %>
-    <div class="tx4 txB row"><%= ob.polyT('orderDetail.contractTab.vendorContractHeading') %></div>
-    <div class="border clrBr clrP clrT rowLg js-jsonVendorContractContainer"></div>
-    <div class="tx4 txB row"><%= ob.polyT('orderDetail.contractTab.buyerContractHeading') %></div>
-    <p class="clrTErr flexVCent"><span class="ion-ios-close-empty tx1 margRSm"></span><%= ob.polyT('orderDetail.contractTab.buyerContractNotAuthentic') %></p>    
-    <div class="border clrBr clrP clrT js-jsonBuyerContractContainer"></div>
-  <% } else { %>
-    <% // vendors contract is forged %>
-    <div class="tx4 txB row"><%= ob.polyT('orderDetail.contractTab.buyerContractHeading') %></div>
-    <div class="border clrBr clrP clrT rowLg js-jsonBuyerContractContainer"></div>
-    <div class="tx4 txB row"><%= ob.polyT('orderDetail.contractTab.vendorContractHeading') %></div>
-    <p class="clrTErr flexVCent"><span class="ion-ios-close-empty tx1 margRSm"></span><%= ob.polyT('orderDetail.contractTab.vendorContractNotAuthentic') %></p>
-    <div class="border clrBr clrP clrT js-jsonVendorContractContainer"></div>
-  <% } %>
-<% } %>
